attribution/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool]

[tool.poetry]
name = "attribution"
version = "0.0.1"
description = "Attribution modelling code using Python and SQL"
authors = [ "pietro greselin <p.greselin@gmail.com>" ]

[tool.poetry.dependencies]
python = "3.9.13"
ChannelAttribution = "2.0.10"
google-cloud-bigquery = ""
openpyxl = "3.1.2"
pandas = "2.0.3"
pyarrow = "14.0.1"
fastparquet = "2023.10.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-clarity = "^1.0.1"
pre-commit = "3.6.0"
mypy = "^0.981"
sqlparse = "^0.4.2"
alembic = "^1.8.1"
notebook = "6.5.4"